Allow onyo-rm to recursively remove directories
`onyo rm` now does remove non-empty directories, recording implicit
operations for its inventory contents.
In addition, this path prepares `onyo rm` for dealing with non-inventory
files (`.onyoignore`) and asset directories.
For non-inventory files, this relies on introducing sort of a dummy
`InventoryOperation` that does not produce a record in the commit
message, but is in-line with other operations with respect to pending
operations and `Inventory.diff()`.

+++ onyo/lib/commands.py
@@ -570,7 +570,10 @@ def onyo_rm(inventory: Inventory,
     for p in paths:
         try:
             inventory.remove_asset(p)
+            is_asset = True
         except NotAnAssetError:
+            is_asset = False
+        if not is_asset or inventory.repo.is_asset_dir(p):
             inventory.remove_directory(p)
 
     ui.print('The following will be deleted:')
